# localstore: remove MultiSeek method and fix concurrent bug