--- a/store/localstore/goleveldb/goleveldb_test.go
+++ b/store/localstore/goleveldb/goleveldb_test.go
@@ -107,22 +107,3 @@ func (s *testSuite) TestSeek(c *C) {
 	c.Assert(k, IsNil)
 	c.Assert(v, IsNil)
 }
-
-func (s *testSuite) TestMultiSeek(c *C) {
-	defer testleak.AfterTest(c)()
-	b := s.db.NewBatch()
-	b.Put([]byte("a"), []byte("1"))
-	b.Put([]byte("b"), []byte("2"))
-	err := s.db.Commit(b)
-	c.Assert(err, IsNil)
-
-	m := s.db.MultiSeek([][]byte{[]byte("z"), []byte("a"), []byte("a1")})
-	c.Assert(m, HasLen, 3)
-	c.Assert(m[0].Err, NotNil)
-	c.Assert(m[1].Err, IsNil)
-	c.Assert(m[1].Key, BytesEquals, []byte("a"))
-	c.Assert(m[1].Value, BytesEquals, []byte("1"))
-	c.Assert(m[2].Err, IsNil)
-	c.Assert(m[2].Key, BytesEquals, []byte("b"))
-	c.Assert(m[2].Value, BytesEquals, []byte("2"))
-}