--- a/store/localstore/kv.go
+++ b/store/localstore/kv.go
@@ -32,50 +32,35 @@ var (
 	_ kv.Storage = (*dbStore)(nil)
 )
 
-type op int
-
-const (
-	opCommit = iota + 1
-)
-
 const (
 	lowerWaterMark = 10 // second
 )
 
-type command struct {
-	op    op
-	txn   *dbTxn
-	args  interface{}
-	reply interface{}
-	done  chan error
-}
-
-type seekReply struct {
-	key   []byte
-	value []byte
-}
-
-type commitReply struct {
-	err error
-}
-
-type seekArgs struct {
-	key []byte
-}
-
-type commitArgs struct {
-}
-
 // Seek searches for the first key in the engine which is >= key in byte order, returns (nil, nil, ErrNotFound)
 // if such key is not found.
-func (s *dbStore) Seek(key []byte) ([]byte, []byte, error) {
-	s.mu.Lock()
-	if s.closed {
+func (s *dbStore) Seek(key []byte, startTS uint64) ([]byte, []byte, error) {
+	for {
+		var conflict bool
+		s.mu.Lock()
+		if s.closed {
+			s.mu.Unlock()
+			return nil, nil, ErrDBClosed
+		}
+		if s.committingTS != 0 && s.committingTS < startTS {
+			// We not sure if we can read the committing value,
+			conflict = true
+		} else {
+			s.wg.Add(1)
+		}
 		s.mu.Unlock()
-		return nil, nil, ErrDBClosed
+		if conflict {
+			// Wait for committing to be finished and try again.
+			time.Sleep(time.Microsecond)
+			continue
+		}
+		// No conflict, safe to seek.
+		break
 	}
-	s.wg.Add(1)
-	s.mu.Unlock()
 	key, val, err := s.db.Seek(key)
 	s.wg.Done()
 	return key, val, err
@@ -86,46 +71,7 @@ func (s *dbStore) CommitTxn(txn *dbTxn) error {
 	if len(txn.lockedKeys) == 0 {
 		return nil
 	}
-	c := &command{
-		op:   opCommit,
-		txn:  txn,
-		args: &commitArgs{},
-		done: make(chan error, 1),
-	}
-
-	s.commandCh <- c
-	err := <-c.done
-	return errors.Trace(err)
-}
-
-func (s *dbStore) scheduler() {
-	closed := false
-
-	segmentIndex := int64(0)
-
-	tick := time.NewTicker(time.Second)
-	defer tick.Stop()
-
-	for {
-		select {
-		case cmd := <-s.commandCh:
-			if closed {
-				cmd.done <- ErrDBClosed
-				continue
-			}
-			switch cmd.op {
-			case opCommit:
-				s.doCommit(cmd)
-			}
-		case <-s.closeCh:
-			closed = true
-			s.wg.Done()
-		case <-tick.C:
-			segmentIndex = segmentIndex % s.recentUpdates.SegmentCount()
-			s.cleanRecentUpdates(segmentIndex)
-			segmentIndex++
-		}
-	}
+	return s.doCommit(txn)
 }
 
 func (s *dbStore) cleanRecentUpdates(segmentIndex int64) {
@@ -169,21 +115,44 @@ func (s *dbStore) tryLock(txn *dbTxn) (err error) {
 	return nil
 }
 
-func (s *dbStore) doCommit(cmd *command) {
-	txn := cmd.txn
-	// Get preWrite version.
-	preWriteVer, err := globalVersionProvider.CurrentVersion()
-	if err != nil {
-		log.Fatal(err)
+func (s *dbStore) doCommit(txn *dbTxn) error {
+	var commitVer kv.Version
+	var err error
+	for {
+		// Atomically get commit version
+		s.mu.Lock()
+		closed := s.closed
+		committing := s.committingTS != 0
+		if !closed && !committing {
+			commitVer, err = globalVersionProvider.CurrentVersion()
+			s.committingTS = commitVer.Ver
+			s.wg.Add(1)
+		}
+		s.mu.Unlock()
+
+		if closed {
+			return ErrDBClosed
+		}
+		if committing {
+			time.Sleep(time.Microsecond)
+			continue
+		}
+		break
 	}
+	defer func() {
+		s.mu.Lock()
+		s.committingTS = 0
+		s.wg.Done()
+		s.mu.Unlock()
+	}()
+	// Here we are sure no concurrent committing happens.
 	err = s.tryLock(txn)
 	if err != nil {
-		cmd.done <- errors.Trace(err)
-		return
+		return errors.Trace(err)
 	}
 	b := s.db.NewBatch()
 	txn.us.WalkBuffer(func(k kv.Key, value []byte) error {
-		mvccKey := MvccEncodeVersionKey(kv.Key(k), preWriteVer)
+		mvccKey := MvccEncodeVersionKey(kv.Key(k), commitVer)
 		if len(value) == 0 { // Deleted marker
 			b.Put(mvccKey, nil)
 			s.compactor.OnDelete(k)
@@ -194,13 +163,27 @@ func (s *dbStore) doCommit(cmd *command) {
 		return nil
 	})
 	err = s.writeBatch(b)
-	// Update commit version.
-	txn.version, err = globalVersionProvider.CurrentVersion()
 	if err != nil {
-		log.Fatal(err)
+		return errors.Trace(err)
 	}
-	s.unLockKeys(txn)
-	cmd.done <- errors.Trace(err)
+	// Update commit version.
+	txn.version = commitVer
+	err = s.unLockKeys(txn)
+	if err != nil {
+		return errors.Trace(err)
+	}
+
+	// Clean recent updates.
+	now := time.Now()
+	if now.Sub(s.lastCleanTime) > time.Second {
+		s.cleanRecentUpdates(s.cleanIdx)
+		s.cleanIdx++
+		if s.cleanIdx == s.recentUpdates.SegmentCount() {
+			s.cleanIdx = 0
+		}
+		s.lastCleanTime = now
+	}
+	return nil
 }
 
 func (s *dbStore) NewBatch() engine.Batch {
@@ -212,18 +195,19 @@ type dbStore struct {
 
 	txns       map[uint64]*dbTxn
 	keysLocked map[string]uint64
-	// TODO: clean up recentUpdates
+
 	recentUpdates *segmentmap.SegmentMap
-	uuid          string
-	path          string
-	compactor     *localstoreCompactor
-	wg            *sync.WaitGroup
+	cleanIdx      int64
+	lastCleanTime time.Time
 
-	commandCh chan *command
-	closeCh   chan struct{}
+	uuid      string
+	path      string
+	compactor *localstoreCompactor
+	wg        sync.WaitGroup
 
-	mu     sync.Mutex
-	closed bool
+	mu           sync.Mutex
+	closed       bool
+	committingTS uint64
 
 	pd localPD
 }
@@ -289,10 +273,7 @@ func (d Driver) Open(path string) (kv.Storage, error) {
 		path:       engineSchema,
 		db:         db,
 		compactor:  newLocalCompactor(localCompactDefaultPolicy, db),
-		commandCh:  make(chan *command, 1000),
 		closed:     false,
-		closeCh:    make(chan struct{}),
-		wg:         &sync.WaitGroup{},
 	}
 	s.recentUpdates, err = segmentmap.NewSegmentMap(100)
 	if err != nil {
@@ -307,8 +288,6 @@ func (d Driver) Open(path string) (kv.Storage, error) {
 	s.pd.SetRegionInfo(infos)
 	mc.cache[engineSchema] = s
 	s.compactor.Start()
-	s.wg.Add(1)
-	go s.scheduler()
 	return s, nil
 }
 
@@ -366,14 +345,9 @@ func (s *dbStore) Close() error {
 		s.mu.Unlock()
 		return ErrDBClosed
 	}
-
 	s.closed = true
 	s.mu.Unlock()
-
-	mc.mu.Lock()
-	defer mc.mu.Unlock()
 	s.compactor.Stop()
-	s.closeCh <- struct{}{}
 	s.wg.Wait()
 	delete(mc.cache, s.path)
 	return s.db.Close()
@@ -383,11 +357,6 @@ func (s *dbStore) writeBatch(b engine.Batch) error {
 	if b.Len() == 0 {
 		return nil
 	}
-
-	if s.closed {
-		return errors.Trace(ErrDBClosed)
-	}
-
 	err := s.db.Commit(b)
 	if err != nil {
 		log.Error(err)
@@ -405,12 +374,11 @@ func (s *dbStore) unLockKeys(txn *dbTxn) error {
 	for k := range txn.lockedKeys {
 		if tid, ok := s.keysLocked[k]; !ok || tid != txn.tid {
 			debug.PrintStack()
-			log.Fatalf("should never happend:%v, %v", tid, txn.tid)
+			return errors.Errorf("should never happend:%v, %v", tid, txn.tid)
 		}
 
 		delete(s.keysLocked, k)
 		s.recentUpdates.Set([]byte(k), txn.version, true)
 	}
-
 	return nil
 }