Before:
```
BenchmarkBasic-4                   50000	     32574 ns/op
BenchmarkTableScan-4         	    1000	   2093847 ns/op
BenchmarkTableLookup-4       	   10000	    112250 ns/op
BenchmarkStringIndexScan-4   	     500	   3506790 ns/op
BenchmarkStringIndexLookup-4 	   10000	    207038 ns/op
BenchmarkIntegerIndexScan-4  	     500	   3444936 ns/op
BenchmarkIntegerIndexLookup-4	   10000	    194365 ns/op
BenchmarkDecimalIndexScan-4  	     300	   4423639 ns/op
BenchmarkDecimalIndexLookup-4	    5000	    231001 ns/op
BenchmarkInsertWithIndex-4   	   20000	    265147 ns/op
BenchmarkInsertNoIndex-4     	   10000	    206556 ns/op
```

After:
```
BenchmarkBasic-4                   50000	     32286 ns/op
BenchmarkTableScan-4         	    1000	   1431793 ns/op
BenchmarkTableLookup-4       	   10000	    105119 ns/op
BenchmarkStringIndexScan-4   	     500	   2420872 ns/op
BenchmarkStringIndexLookup-4 	   10000	    181779 ns/op
BenchmarkIntegerIndexScan-4  	    1000	   2436715 ns/op
BenchmarkIntegerIndexLookup-4	   10000	    185647 ns/op
BenchmarkDecimalIndexScan-4  	     500	   3067935 ns/op
BenchmarkDecimalIndexLookup-4	    5000	    211426 ns/op
BenchmarkInsertWithIndex-4   	   20000	    138176 ns/op
BenchmarkInsertNoIndex-4     	   10000	    144457 ns/op
```