@coocood 2016-04-26
MultiSeek can only improve concurrent performance but for local store, we just use it to run tests.