@qiuyesuifeng 2016-04-26
I remember that using multiSeek can also resolve a transaction problem, but i am not sure for that.