@coocood 2016-04-28
Removed local store scheduler.
@shenli @disksing PTAL